// LinkERP/ClientApp/src/app/shared/enums/enum-generictables.ts
export enum GenricTableIDByName {
  LBS_INV_UnitOfMeasure = "79C0F939-081C-4C7E-83F5-B4DFB2F04E66",
  LBS_INV_Categories = "CA973CE8-8CDD-4C3B-81A5-C09257C56A01",
  LBS_INV_StockTakeCycle = "3A65DA85-1441-4DD8-B6B7-3575EBF99BDE",
  LBS_SYS_NoteTypes = "4FE7E4BD-CA18-4EB7-96F1-49D3F58208DB",
  LBS_SOP_TenderTypeGroups = "D09C17DB-5A26-4BE3-8203-4517648D70F7",
  LBS_SOP_TenderFieldType = "5E80B3F3-7592-4CEB-9A9C-FB72A8684D48",
  LBS_SOP_QuotationStatus = "B63AD2D6-6C2F-41C2-B956-4668E7A9EFE7",
  LBS_SOP_TenderTypesLookupList = "5E80B3F3-7592-4CEB-9A9C-FB72A8684D49",
  LBS_INV_ProductGroup = "3D3C2595-8FF0-419B-A42C-0E5CA0D22852",
  LBS_INV_WarrantyFrequency = "16A9F43E-055B-467E-8D6B-F7C86E4A6403",
  LBS_INV_ProductGroup2 = "59D25ED3-B5F8-44A2-B549-56DEB9CD2AC1",
  LBS_INV_ProductGroup3 = "81A64179-4C73-47FB-A700-8E9D9D9B45A7",
  LBS_SOP_Instruction_Types = "7F0228D9-6328-4425-B63E-B1F9F2D79614",
  LBS_SOP_GenderType = "02191B92-FE5C-4ADB-90B1-EDD4D10DA2F7",
}  